+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Rustdoc for `stuck::main` and `stuck::test`.
+
 ## [0.1.3] - 2022-04-07
 ### Added
 - Add proc macros to bootstrap runtime for main and test