+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.1.5] - 2022-04-09
 ### Added
 - `Sender::try_send` to send without blocking current execution.
 - Network io poller and tcp support.
@@ -43,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
+[0.1.5]: https://github.com/kezhuw/stuck/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/kezhuw/stuck/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/kezhuw/stuck/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/kezhuw/stuck/compare/v0.1.1...v0.1.2